requirements.txt
pyqt6-tools==6.4.2.3.3
PyQt6==6.5.0
PyQt6-Qt6==6.5.0
PySide6~=6.5.2
pyinstaller==6.2.0
Pillow~=10.1.0